Move BatchWriter to kvstore package
