+++ objectstorage/object_storage.go
@@ -917,7 +917,7 @@ func (objectStorage *ObjectStorage) flush() {
 	// force release the collected objects
 	for j := 0; j < i; j++ {
 		if consumers := atomic.AddInt32(&(cachedObjects[j].consumers), -1); consumers == 0 {
-			objectStorage.options.batchedWriterInstance.batchWrite(cachedObjects[j])
+			objectStorage.options.batchedWriterInstance.Enqueue(cachedObjects[j])
 		}
 	}
